requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
Pillow
openai
python-dotenv
